requirements.txt
hydra-core==1.3.2
opencv-python==4.9.0.80
requests==2.31.0
scikit-learn==1.2.2
psycopg2-binary==2.9.9
vidgear==0.3.2
Flask==3.0.3
pandas==2.2.2
clickhouse_driver
ultralytics==8.3.10
kafka-python==2.0.2
hydra-core==1.3.2
minio==7.2.8